requirements.txt
django==3.2
Pillow>=8.1.2